refactor elements phase 1
