+++ src/components/elements/divider.js
@@ -1,8 +1,10 @@
 import styled from 'styled-components'
 
-export const Divider = styled.hr`
+const Divider = styled.hr`
     border: none;
     width: 100%;
     height: 1px;
     background-color: var(--color-grey-2);
 `
+
+export default Divider
