+++ src/components/elements/index.js
@@ -1,0 +1,34 @@
+import AutoCarousel from './auto-carousel'
+import { Card, CardChildren } from './card'
+import Carousel from './carousel'
+import Divider from './divider'
+import DotLoader from './dot-loader'
+import Dropdown from './dropdown'
+import Hero from './hero'
+import Image from './image'
+import Link from './link'
+import LinearLoader from './loader'
+import Modal from './modal'
+import OffCanvas from './off-canvas-menu'
+import SideTab from './side-tab'
+import { Text, LinkText, Header } from './typography'
+
+export {
+    AutoCarousel,
+    Card,
+    CardChildren,
+    Carousel,
+    Divider,
+    DotLoader,
+    Dropdown,
+    Hero,
+    Image,
+    Link,
+    LinearLoader,
+    Modal,
+    OffCanvas,
+    SideTab,
+    Text,
+    LinkText,
+    Header,
+}
