+++ src/components/layout/footer.js
@@ -8,7 +8,7 @@ import {
 } from '../localization'
 import { Container } from '../containers'
 import { Header, Text } from '../elements/typography.js'
-import { StyledLink } from '../elements/link'
+import StyledLink from '../elements/link'
 import { isProduction } from 'common/websocket/config'
 import device from 'themes/device'
 // Icons
