+++ src/pages/keep-safe/index.js
@@ -16,8 +16,8 @@ import Button from 'components/form/button'
 import Modal, { useModal } from 'components/elements/modal'
 import SignupModal from 'components/custom/signup-modal'
 import { Card } from 'components/elements/card.js'
-import { Divider } from 'components/elements/divider'
-import { StyledLink } from 'components/elements/link'
+import Divider from 'components/elements/divider'
+import StyledLink from 'components/elements/link'
 import { Header, Text } from 'components/elements/typography.js'
 import { localize, WithIntl, Localize } from 'components/localization'
