+++ src/pages/terms-and-conditions/_styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import { Header, Text } from 'components/elements/typography'
-import { StyledLink } from 'components/elements/link'
+import StyledLink from 'components/elements/link'
 import device from 'themes/device'
 
 export const StyledSection = styled.section`
